refactor: organize source code by concern
- Moved files copied from 3rd party to lib/
- Moved all protobuf files to src/protobuf
