+++ lib/advertisement.py
