+++ src/protobuf/protos/assert_location.proto
