+++ src/protobuf/protos/diagnostics.proto
