+++ src/protobuf/protos/wifi_remove.proto
